+++ shuup/front/__init__.py
@@ -34,6 +34,7 @@ class ShuupFrontAppConfig(AppConfig):
             "shuup.front.notify_events:RefundCreated",
         ],
         "front_extend_product_list_form": [
+            "shuup.front.forms.product_list_modifiers.CategoryProductListFilter",
             "shuup.front.forms.product_list_modifiers.SortProductListByName",
             "shuup.front.forms.product_list_modifiers.SortProductListByPrice",
             "shuup.front.forms.product_list_modifiers.ManufacturerProductListFilter",